Bigger icons
I just made the icons fill all the space available because I thought they were too small with the margin that was left previously. I think this will end up being personal preference whether you implement it or not, but here it is in case you like it. Maybe in the future I will try to make a new icon.